Moved things around as requested in PR
